LayoutTests/svg/animations/animate-path-animation-lL-vV-hH-inverse-expected.txt

SVG 1.1 dynamic animation tests

Test path animation where coordinate modes of start and end differ. You should see PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is 'M'
PASS path.pathSegList.getItem(0).x is 30
PASS path.pathSegList.getItem(0).y is 30
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is 'l'
PASS path.pathSegList.getItem(1).x is -60
PASS path.pathSegList.getItem(1).y is -30
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is 'v'
PASS path.pathSegList.getItem(2).y is -30
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is 'h'
PASS path.pathSegList.getItem(3).x is 30
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is 'M'
PASS path.pathSegList.getItem(0).x is 15
PASS path.pathSegList.getItem(0).y is 15
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is 'l'
PASS path.pathSegList.getItem(1).x is -30
PASS path.pathSegList.getItem(1).y is -15
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is 'v'
PASS path.pathSegList.getItem(2).y is -15
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is 'h'
PASS path.pathSegList.getItem(3).x is 15
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is 'M'
PASS path.pathSegList.getItem(0).x is -15
PASS path.pathSegList.getItem(0).y is -15
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is 'L'
PASS path.pathSegList.getItem(1).x is 15
PASS path.pathSegList.getItem(1).y is 0
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is 'V'
PASS path.pathSegList.getItem(2).y is 15
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is 'H'
PASS path.pathSegList.getItem(3).x is 0
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is 'M'
PASS path.pathSegList.getItem(0).x is -30
PASS path.pathSegList.getItem(0).y is -30
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is 'L'
PASS path.pathSegList.getItem(1).x is 30
PASS path.pathSegList.getItem(1).y is 0
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is 'V'
PASS path.pathSegList.getItem(2).y is 30
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is 'H'
PASS path.pathSegList.getItem(3).x is 0
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is 'M'
PASS path.pathSegList.getItem(0).x is 30
PASS path.pathSegList.getItem(0).y is 30
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is 'l'
PASS path.pathSegList.getItem(1).x is -60
PASS path.pathSegList.getItem(1).y is -30
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is 'v'
PASS path.pathSegList.getItem(2).y is -30
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is 'h'
PASS path.pathSegList.getItem(3).x is 30
PASS successfullyParsed is true

TEST COMPLETE
